Refactored to add mobile styles and cleaned up file organization
