+++ voyage-45/src/app/layout.js
@@ -1,4 +1,4 @@
-import './globals.css'
+import '../components/styles/globals.css'
 import { Inter } from 'next/font/google'
 
 const inter = Inter({ subsets: ['latin'] })
