+++ voyage-45/src/components/StrikesByComp.js
